Initial check-in. LRU cache
Attempted a solution that used a dictionary and a variable that kept a
reference to the least used value, realized that I need another
structure that allows me to keep a 'history' of changes to backtrack as
much as necesary when adding new elements and removeing least used ones.

Probably an ordered dictionary but I'm not sure if it will be O(1) for
all operations in that case. Reasearching possible solutions.